requirements.txt
aiogram>=3.4,<4
openai>=1.0
pyyaml>=6.0
jinja2>=3.1
python-dotenv>=1.0
